Debug added; beacon() is not being called
Serial.print statements were added to tell which part of the code is being executed. The first line of beacon(), now such a statement, is not being called.

A few small changes have been added to beacon() and the test code in futile fix attempts.

The hard link did not seem to work as expected; changes from Morse_Code.h did not appear when accessing the link.