more efuse behavior changes
Okay, that was deeper than I thought.  efuse.overcurrent() now is a void function and just sends out the message to avionics if the threshold is exceeded.
I might add in a debug message later.
That change necessitated changes commands.cpp, beacon.cpp and testing_support.cpp.  There is no direct function for "is there an overcurrent condition?", but you can always just ask for the current and figure it out yourself.  So there.