several changes all at once
love it when i do this...
modified overcurrent to return a 1 if an overcurrent has occured.  This is so it can be saved off.   Also modified the function so it also checks the fault pin and if that's set it also reports the overcurrent.

RadioHW was modified to pass the global fault parameter

commands.cpp modified to pass the global fault parameter

commands2.cpp was modified to make processcmdbuff to return a boolean.   processcommand was modified to pass the global fault parameter.  status was modified too.  halt was renamed to reset.

e2ecommands_metro.ino modified to add the fault parameter and to pass it where needed.  Switched to commands2 library and modifed it to use the class functions.

efuse.overcurrent now returns true if there's an overcurrent or if the fault line is low.