Update packetfinder.h
simplified the packet length function, but this is not yet tested, it just compiles.
It now only returns a value if there's more than a minimal length command in the buffer, data should always be larger than 3 bytes because of the tncattach header.
I'm also no longer checking for the next 0xC0 because it might not exist (?) need to make sure all of the packet is being cleared.