tracking down a "not a bug"
modified all references for CircularBuffer to .hpp per the compiler warnings.
removed old commands.h and commands.cpp
corrected indexing on body processing in commands2.cpp
added #define for serial buffer size (this has not been tested)
put define in to try to make circular buffer interrupt safe, but that may not be needed.  Issue I was chasing down appears to be overrunning the serial receive buffer...especially since the loop is taking about 2.2mS to execute.  If a second packet comes in during that time, it's easy to overrun the 64 byte default serial buffer.  So my thought is to try to increase that buffer to a point where a whole packet or two can be buffered and then drawn down.  Fundamentally, you're always limited by the average data rate, which can only approach but not exceed the over the air baud rate.
There's some leftover debug code left in packetfinder.h that probably can be removed later on.