abstraction and serial0 dup
added code to dup the comms on Serial0 to Serial.  This is to facilitate testing after the satellite is sealed up

Also restructured the beacon to move dit and dah back into the beacon code and replaced them in the radio class with a single 'key' which takes an integer of the number of 'chips' that you want to send (multiple of constants::bit_time).

Also restructured the radio configuration code that was in Setup() and moved it into radio.begin().

deleted some debug messages too along the way.

THIS IS ALL UNTESTED...be warned!