changed sweep result format
I eliminated the arrays that were holding the results.  It now just prints to the debug port at each frequency.
I also added a delay between rssi readings.  Set it arbitrarily to 10mS.  But that does seem to have improved the results.  It now detects the other transmitter when it's outputting CW.