ax_structures
I created the file, "ax_structures.h", which holds the common set of structures used across the various ax library files ( in particular ax.h, ax_hw.h, and ax_params.h).  This broke the circular dependency between the files.  ax_structures is now an include in each of those files.  This allows the removal of the extra includes in ax.cpp.

This has been compiled, but is not yet tested.