non-volatile frequency and reset
I've added the code to support storing the operating frequency in the internal flash of the SAMD21.  This is untested.  It will reset to the default (in constants.cpp) if the code is reloaded.  Otherwise, it should retain the value.  It only retains when told the same frequency twice in a row.  Otherwise, it's a temporary change.
I also modified the reset command to be a hard reset.  This is done via the external watchdog.