fixes to Tx sweep
Support for tx sweeps over larger ranges was added.
Wider sweeps require the VCO to be re-ranged.  This had unintended consequences when operating in wire mode.  Also, the ranging function leaves Synth B selected, so that made me create two new functions to select the appropriate synth depending on the tx/rx mode.
Changed back to the all-internal VCO, but may be able to change it back.  Still not sure what is the advantage of the external inductor.