fixes to make receiver sweep work, adding in silversat ground
Fixed issue with wrong VCO getting selected for receiver sweep
Added in the Silversat Ground Radio.  This needs the new variant (Silversat Ground) in order to work.  The ports were swapped (Avionics is Payload, Payload is Avionics), and we need to disable the eFuse messages (there's no efuse in the ground station).