add max current & averaging fix
added the ability to track the max current in the efuse class.  The value can be cleared using clear_max_current().
I also revised the way the integrated rssi was being calculated to make it consistent with an earlier function that does the same thing.  I was calculating a running average, when all I really needed to calculate was the average at the end of a period.
Also cleaned up the variable types a bit.  Shouldn't really affect the calculation, but it's more specific.