fixes from bad merge
All the back and forth preceding this commit was an attempt to avoid what this commit had to do.  That is, I had to reapply some changes that got wiped out when I did a bad merge/revert merge.  I still need to text the code.
This puts back the variant for the ground radio board (which got dropped by the bad merge)
This also *hopefully* fixes the issue where the sweep was reporting an RSSI of 0 (it wasn't before the merge...I think it's because the chip is off).
I still need to pull the correct boards.txt from my development computer.  There also may be other changes to the variants files there, but I'm just going to do a blanket copy.