multiple changes
changed current measure to integers to limit the size of the status message better
added length checks on all commands.  added some parameter checks on individual commands.  It would probably be good to make sure conversions from string to integer return a number.
fixed error in implementation of adjust power command.  I had it as a 3 byte value, when it should have been 1 byte, with power levels 0-10.  I had to also fix this in the GUI.
Still Testing.