radio class additional functions
added some additional functions to radio class.  I'm trying to eliminate all direct AX calls outside of the main libraries.

Propagated those changes to the files as needed.

Deprecated the toggle frequency command.  I don't feel like trying to make it work ;)

Added a lot of debug messages to try to eliminate collisions.  It's still colliding on occasion and Im not too sure why yet.  Changed the CCA to be the max of five sequential measurements spaced about 1mS apart.  That was based on the assumption that the rssi wasn't consistently reporting.  Also tried radiostate, but that's not of much use when it's receiving.  It doesn't seem to accurately indicate when it gets a valid preamble.  Really would be nice if it did...or I don't know how to get that working.
Starting fixing some comments, but there's a lot to fix...