PA current draw mitigation
Testing has shown that it's possible for the PA to enter into a high current state if the PA is enabled without RF input.

Now the PA does not get enabled until Radio.transmit is executed.  The next issue is determining when to turn it back off.  I enabled the interrupt for the radio event (transmit or receive done), and rather than blocking, I'm waiting for the interrupt to trigger.  When that triggers, it always turns the PA off (so I don't have to determine if the interrupt triggered from a tx or rx event. )
I may want to disable interrupts while in RECEIVE and enable them while in TRANSMIT