increased chunk size, and il2p fixes
Increased the chunk size before splitting to 239 (from 200) to give us a little more space to make up for the IL2P layer.  Packets in excess of 239 are getting garbled on the receiver side (not sure why).
This version adds the interrupt logic to turn off the PA when a packet transmission is complete.  The PA now turns on when the data is transmitted.
Preambles increased to 16, but probably not necessary in the long run.  Needs testing.
Dropped TXBUFFSIZE to 512.  Probably could drop that further, but no less than 256.
Added board reset indicator.  Will be interesting to see if watchdog ever fires.
Added lots of VERBOSE debug messages, but left the logging level at WARNING
Fixed copyToArray error...nice function, but it leaves the data in the buffer and you end up having to empty it anyway.
Adjusted some of the variable types in the eFuse to get the output value to be correct (although I thought I tested that already...)
Added radio function to print out the modulation parameters.
Fixed the formatting on all the register logging in testing_support.cpp.