stats and fixes for radiostate
I've added a Stats structure which contains all the high water marks for each buffer and the process times for all the main sections of code.
The corresponding command to get the Stats is also included.
I also have fixed the issue with 'radiostate' .  It's the actual state of the radio state machine, so it returns to different states depending on the power state (FULLTX vs. FULLRX)