candidate version 1.12
update threshold for receive overcurrent
turned on RSSI and RF_OFFSET meta data.  These are output via the debug port.
changed command order in cw mode to ensure PA is not turned on until RF is, and vice versa.
Changed radio::key() to do much the same.  RF first, then PA, but with a 10mS delay.  Delay is not optimized, but current monitor output looks much better.
Did the same in radio::transmit(), but did not add the delay.  Doesn't appear necessary based on current monitor output.  Still have a step change in current when RF turns off, but it's not that big.